+++ frontend/package.json
@@ -12,8 +12,10 @@
     "leaflet": "^1.7.1",
     "package.json": "^2.0.1",
     "react": "^17.0.2",
+    "react-bootstrap": "^2.0.4",
     "react-dom": "^17.0.2",
     "react-image-gallery": "^1.2.7",
+    "react-router-dom": "^6.2.1",
     "react-script": "^2.0.5",
     "react-scripts": "5.0.0",
     "reactstrap": "^8.9.0",
